fix: refactor test
